Created the GarageService class that contains all person, productionOrder and vehicle methods. Removed the old service classes.
