+++ src/main/java/com/daitangroup/api/controller/PersonController.java
@@ -1,9 +1,7 @@
 package com.daitangroup.api.controller;
 
 import com.daitangroup.api.model.Person;
-import com.daitangroup.api.model.ProductionOrder;
-import com.daitangroup.api.model.Vehicle;
-import com.daitangroup.api.services.PersonService;
+import com.daitangroup.api.services.GarageService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,12 +14,12 @@ import java.util.Optional;
 public class PersonController {
 
     @Autowired
-    private PersonService personService;
+    private GarageService garageService;
 
     @RequestMapping(method = RequestMethod.GET, value = "/{id}")
     public ResponseEntity<Person> findPersonById(@PathVariable Long id) {
 
-        Optional<Person> personResult = personService.findPersonById(id);
+        Optional<Person> personResult = garageService.findPersonById(id);
 
         if (personResult.isPresent()) {
             return new ResponseEntity<>(personResult.get(), HttpStatus.OK);
@@ -33,7 +31,7 @@ public class PersonController {
     @RequestMapping(method = RequestMethod.GET)
     public ResponseEntity<Iterable> findAllPersons() {
 
-        Iterable<Person> personsResult = personService.findAllPersons();
+        Iterable<Person> personsResult = garageService.findAllPersons();
 
         if (personsResult.iterator().hasNext()) {
             return new ResponseEntity<>(personsResult, HttpStatus.OK);
@@ -45,7 +43,7 @@ public class PersonController {
     @RequestMapping(method = RequestMethod.GET, value = "/cpf/{cpf}")
     public ResponseEntity<Person> findPersonByCpf(@PathVariable String cpf) {
 
-        Optional<Person> personResult = personService.findPersonByCpf(cpf);
+        Optional<Person> personResult = garageService.findPersonByCpf(cpf);
 
         if (personResult.isPresent()) {
             return new ResponseEntity<>(personResult.get(), HttpStatus.OK);
@@ -57,7 +55,7 @@ public class PersonController {
     @RequestMapping(method = RequestMethod.POST, value = "/create")
     public ResponseEntity<Person> createPerson(@RequestBody Person person) {
 
-        Person personSaved = personService.createPerson(person);
+        Person personSaved = garageService.createPerson(person);
 
         if(personSaved != null){
             return new ResponseEntity<>(person, HttpStatus.OK);
@@ -69,7 +67,7 @@ public class PersonController {
     @RequestMapping(method = RequestMethod.POST, value = "/associatevehicle")
     public ResponseEntity<Person> addVehicleToPerson(@RequestParam("personid") long personid, @RequestParam("vehicleid") long vehicleid) {
 
-        Person person = personService.addVehicleToPerson(personid, vehicleid);
+        Person person = garageService.addVehicleToPerson(personid, vehicleid);
 
         if (person != null) {
             return new ResponseEntity<>(person, HttpStatus.OK);
@@ -82,7 +80,7 @@ public class PersonController {
     public ResponseEntity<Person> updatePerson(@PathVariable long id, @RequestBody Person person) {
 
         person.setId(id);
-        Person personUpdated = personService.updatePerson(person);
+        Person personUpdated = garageService.updatePerson(person);
 
         if(person != null) {
             return new ResponseEntity<>(personUpdated, HttpStatus.OK);
@@ -94,7 +92,7 @@ public class PersonController {
     @RequestMapping(method = RequestMethod.DELETE, value = "/{id}")
     public ResponseEntity<Person> deletePerson(@PathVariable long id) {
 
-        Person personDeleted = personService.deletePerson(id);
+        Person personDeleted = garageService.deletePerson(id);
 
         if(personDeleted != null){
             return new ResponseEntity<>(personDeleted, HttpStatus.OK);
@@ -106,7 +104,7 @@ public class PersonController {
     @RequestMapping(method = RequestMethod.DELETE, value = "/deleteall")
     public ResponseEntity<String> deleteAllPersons() {
 
-        personService.deleteAllPersons();
+        garageService.deleteAllPersons();
 
         return new ResponseEntity<String>("All persons deleted",HttpStatus.OK);
     }
